test(fingerprint): split policy-version tests to own file (FR-006 cap)
cohort-config-fingerprint.test.ts grew past the 500-line cap with the #1466 v2
updates; the policy-version unit tests move to redaction-policy-version-bump.test.ts
(which also gains the PII-settings coverage).
