+++ tests/unit/memory/cohort-config-fingerprint.test.ts
@@ -34,49 +34,6 @@ function makeTmp(prefix: string): string {
 }
 
 describe('#1850 cohort config fingerprint inputs (acceptance #10, #13)', () => {
-	test('computeRedactionPolicyVersion is deterministic', () => {
-		const a = computeRedactionPolicyVersion({ rejectDurableSecrets: true });
-		const b = computeRedactionPolicyVersion({ rejectDurableSecrets: true });
-		expect(a).toBe(b);
-	});
-
-	test('rejectDurableSecrets=true differs from false', () => {
-		const withReject = computeRedactionPolicyVersion({
-			rejectDurableSecrets: true,
-		});
-		const withoutReject = computeRedactionPolicyVersion({
-			rejectDurableSecrets: false,
-		});
-		expect(withReject).not.toBe(withoutReject);
-		expect(withReject).toBeGreaterThan(withoutReject);
-	});
-
-	test('#1466 PII settings change the policy version', () => {
-		const base = computeRedactionPolicyVersion({ rejectDurableSecrets: true });
-		const withDetect = computeRedactionPolicyVersion({
-			rejectDurableSecrets: true,
-			detectPii: true,
-		});
-		const withReject = computeRedactionPolicyVersion({
-			rejectDurableSecrets: true,
-			rejectDurablePii: true,
-		});
-		const withNer = computeRedactionPolicyVersion({
-			rejectDurableSecrets: true,
-			piiDetector: 'ner',
-		});
-		expect(withDetect).not.toBe(base);
-		expect(withReject).not.toBe(base);
-		expect(withNer).not.toBe(base);
-		expect(withDetect).not.toBe(withReject);
-	});
-
-	test('version is a positive integer', () => {
-		const v = computeRedactionPolicyVersion(true);
-		expect(Number.isInteger(v)).toBe(true);
-		expect(v).toBeGreaterThan(0);
-	});
-
 	test('computeMemoryCohortFingerprint is deterministic + 12 hex chars', () => {
 		const input = buildMemoryCohortFingerprintInput(DEFAULT_MEMORY_CONFIG);
 		const a = computeMemoryCohortFingerprint(input);
@@ -117,13 +74,10 @@ describe('#1850 cohort config fingerprint inputs (acceptance #10, #13)', () => {
 	 * change to `stableCanonicalStringify` and is unchanged after it — empirical
 	 * proof that the change did not invalidate persisted fingerprints.
 	 *
-	 * #1466: version 2 of the algorithm — the digest changes for unchanged
-	 * user config because REDACTION_POLICY_SALT went 1→2 (PII policy fields
-	 * joined computeRedactionPolicyVersion) and a secret-pattern family was
-	 * added. Per the FINGERPRINT_ALGORITHM_VERSION bump rule this is an
-	 * algorithm bump: legacy v1 cohort files fail open via the version gate
-	 * with a "re-run /swarm memory link" advisory instead of a false
-	 * config-mismatch throw. The golden below pins the v2 algorithm.
+	 * #1466: version 2 — the digest changed for unchanged user config
+	 * (REDACTION_POLICY_SALT 1→2 with the PII policy fields + a new secret
+	 * family), so per the bump rule legacy v1 files fail open via the version
+	 * gate with a re-link advisory. The golden below pins the v2 algorithm.
 	 */
 	test('golden digest pins the fingerprint algorithm (version 2)', () => {
 		const golden: MemoryCohortFingerprintInput = {
@@ -170,17 +124,13 @@ describe('#2062 F-012 classifyStoredFingerprintAlgorithmVersion', () => {
 	test('absent version is legacy v1, NOT the current version', () => {
 		// The core bug. With an explicit current an absent field must resolve
 		// to 1 and therefore MISMATCH — not silently equal current and
-		// byte-compare.
+		// byte-compare. (#1466 made the real current version 2, so the no-arg
+		// call now also mismatches: legacy files fail open with the advisory.)
 		expect(classifyStoredFingerprintAlgorithmVersion(undefined, 2)).toEqual({
 			status: 'mismatch',
 			storedVersion: 1,
 			currentVersion: 2,
 		});
-		// #1466 made the real current version 2 (REDACTION_POLICY_SALT bump),
-		// so the no-arg call — which compares against the CURRENT version —
-		// now also reports a mismatch for absent fields: legacy v1 files fail
-		// open with the re-link advisory instead of byte-comparing against a
-		// v2 digest.
 		expect(classifyStoredFingerprintAlgorithmVersion(undefined)).toEqual({
 			status: 'mismatch',
 			storedVersion: 1,
@@ -258,10 +208,9 @@ describe('#1850 SQLite provider fingerprint enforcement (acceptance #10 fail-clo
 		dirs.push(cohortRoot);
 		const dir = makeTmp('fp-worktree-mismatch-');
 		dirs.push(dir);
-		// Write a cohort-config.json carrying the CURRENT algorithm version
-		// with a DELIBERATELY wrong fingerprint. (#1466: without
-		// algorithm_version the file is legacy v1, which now takes the
-		// version-mismatch fail-open path — see the legacy tests below.)
+		// #1466: carries the CURRENT algorithm version with a DELIBERATELY
+		// wrong fingerprint (without algorithm_version it is legacy v1 and
+		// takes the fail-open path — see the legacy tests below).
 		writeFileSync(
 			path.join(cohortRoot, 'memory-cohort-config.json'),
 			JSON.stringify({
@@ -275,8 +224,6 @@ describe('#1850 SQLite provider fingerprint enforcement (acceptance #10 fail-clo
 			DEFAULT_MEMORY_CONFIG,
 			cohortRoot,
 		);
-		// initialize MUST throw — the stored fingerprint does not match what
-		// this worktree's config computes.
 		await expect(provider.initialize()).rejects.toThrow(/fingerprint mismatch/);
 	});
 
@@ -410,19 +357,16 @@ describe('#2062 F-012 cohort config algorithm_version handling', () => {
 	});
 
 	test('legacy file with no algorithm_version fails open after the #1466 v2 bump (no strand)', async () => {
-		// Pre-#1466 this test asserted a fail-closed throw for absent-version
-		// files. FINGERPRINT_ALGORITHM_VERSION is now 2, so an absent version
-		// means legacy v1: digests are not comparable, and the sanctioned
-		// upgrade path is fail-open with the re-link advisory (memory is
-		// never stranded over an algorithm bump alone).
+		// Absent version now means legacy v1 under algorithm v2: digests are
+		// not comparable, so the sanctioned path is fail-open with the
+		// re-link advisory — never a strand over an algorithm bump.
 		const cohortRoot = writeStoredConfig({ fingerprint: 'deadbeefdead' });
 		const provider = new SQLiteMemoryProvider(
 			worktree(),
 			DEFAULT_MEMORY_CONFIG,
 			cohortRoot,
 		);
-		// Opens (warns) instead of throwing.
-		await provider.initialize();
+		await provider.initialize(); // opens (warns) instead of throwing
 		provider.close();
 	});
